+++ test/animated.test.js
@@ -6,7 +6,7 @@ import { create as r } from 'react-test-renderer';
 
 import s from '../src';
 
-test('works with animated components', () => {
+it('works with animated components', () => {
   const Foo = s(Animated.View)({ flex: 1 });
   const foo = r(<Foo />).toJSON();
   expect(foo.props.style).toEqual({ flex: 1 });