+++ test/props.test.js
@@ -4,7 +4,7 @@ import { create as r } from 'react-test-renderer';
 
 import s from '../src';
 
-test('accepts a style prop', () => {
+it('accepts a style prop', () => {
   const Foo = s.View({ margin: 10 });
   const foo = r(<Foo style={{ flex: 1 }} />).toJSON();
   expect(foo.props.style).toEqual({ margin: 10, flex: 1 });