+++ test/ref.test.js
@@ -4,7 +4,7 @@ import { create as r } from 'react-test-renderer';
 
 import s from '../src';
 
-test('forwards ref to wrapped component', () => {
+it('forwards ref to wrapped component', () => {
   const Foo = s.Touchable({ margin: 10 });
   let barRef;
   class Bar extends React.Component {
@@ -23,7 +23,7 @@ test('forwards ref to wrapped component', () => {
   expect(bar).toMatchSnapshot();
 });
 
-test('forwards ref created with React.createRef() to wrapped component', () => {
+it('forwards ref created with React.createRef() to wrapped component', () => {
   const Foo = s.Touchable({ margin: 10 });
   let barRef;
   class Bar extends React.Component {