+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.0.5
+
+- Remove 'prop-types' from dependencies.
+- Display names are now on the format `styled(Foo)` instead of `StyledFoo`.
+- Decouple from `react-native`, it can be used with React DOM.
+
 # 0.0.4
 
-* Initial release.
+- Initial release.